requirements.txt
fastapi==0.89.1
uvicorn==0.20.0
python-multipart
coverage==6.3.2
httpx
twilio
openai-whisper